AI_bug_bountyhunter/requirements.txt
requests
colorama
beautifulsoup4
tqdm
flask
streamlit
python-dotenv
